compiles, some functionality
The source base is compiling. Signaling of updated properties that effect rendering is not in place. Some functionality of the cairo graphics suite must be placed inside.  Recheck the design philosophy idioms to make sure all objects are named appropiately. For example, the use of coordinates as well as move_to. Perhaps having an object named clipped may simplify.