source back - not compiling
Corrected several more design and compiler errors. As well, ensuring maintainability of the code base through documentation and easier to read code. Currently there is an issue with the design that need correction. The inheritance of invoke and emit from multiple base classes causes an ambiguous name. I thought that by deriving the methods of empty abstraction from a base class polymorphic_overloads_t would resolve those.

However the architecture of inheriting public constructors with the using A::A syntax works well. However it seems that the data members in the class are not befriended correctly. The class storage emitters and class drawing functions (MACRO DEFINED) are declared friends of their counter parts. However data seems not accessible to the compiler. Perhaps the definition is backwards.