source back - not compiling
The public method overloading of the emitters are not compiling. Perhaps a study on the technique will resolve more cleanly. Currently the invoke method is placed into the definition with the MACRO and calls the associated storage type. However, the emitters have multiple prototypes that may or may not be defined. Currently it appears that the emit(PangoLayout*) may be resolving as their are not complaints about it.