source back - not compiling
there appears to be 1 compiler issue. However, there are numerous warning to resolve. The inheritance model is not as expected and hence the ctor order is complained about. As well, hidden methods of overloaded functions. Need to evaluate the complete design to see if other higher level changes should occur with the template design. Perhaps the design could be refined even more.